Change pattern for updating project property
Previous it was a delete and then create, there is no upsert available.
But that wasn't working due to a bug somewhere in the graphql lib, where
the delete didn't play with with how we have authorization set up on the
project level. Debugging wasn't yielding much fruit so I updated it to
create the property explicitly from the form and only update it after.