Add a postinstall path to remove a tab character
This tab character is somehow being written literally into the query and
breaking it, only in production. "\tDETACH" appears as literals which is
of course not a valid keyword. Really at a loss for how that could
happen but going to try and validate that this is what it is.